requirements.txt
fastapi
uvicorn
sqlalchemy[asyncio]
asyncpg
joblib
pandas
pydantic
python-dotenv
passlib
pyjwt
scikit-learn
groq
bcrypt==3.2.0
scikit-surprise
faiss-cpu